+++ sdk/search/azure-search-documents/tests/async_tests/test_buffered_sender_async.py
@@ -20,7 +20,7 @@ class TestSearchBatchingClientAsync(object):
     async def test_search_indexing_buffered_sender_kwargs(self):
         async with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL, window=100) as client:
             assert client._batch_action_count == 512
-            assert client._max_retries == 3
+            assert client._max_retries_per_action == 3
             assert client._auto_flush_interval == 60
             assert client._auto_flush